+++ opentmf-api-clients-common/src/main/java/org/opentmf/api/client/common/model/Scope.java
@@ -10,6 +10,7 @@ public enum Scope {
   GET("get"),
   LIST("list"),
   POST("post"),
+  PUT("put"),
   PATCH("patch"),
   DELETE("delete");